v8.4: Continuation of r16347 on filtering local definitions. Refined the "choose less dependent" constraint-solving heuristic so that it is not disturbed by local definitions.
This is a quick fix. A deeper analysis of the structure of constraints
of the form ?x[args] = y, determining if variable y can itself be a
local def or not, and whether args can be let-ins aliasing other
variables, would allow to know if the fix needs to be refined further.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/coq/branches/v8.4@16377 85f007b7-540e-0410-9357-904b9bb8a0f7